Remove of the VENTouchLock.
