+++ Trust/Lock/Lock.swift
@@ -4,10 +4,12 @@ import UIKit
 import SSKeychain
 
 class Lock {
-    struct Keys {
+    private struct Keys {
         static let service = "trust.lock"
         static let account = "trust.account"
     }
+    private let standardDefaults = UserDefaults.standard
+    private let passcodeAttempts = "passcodeAttempts"
     func isPasscodeSet() -> Bool {
         return currentPasscode() != nil
     }
@@ -22,5 +24,17 @@ class Lock {
     }
     func deletePasscode() {
         SSKeychain.deletePassword(forService: Keys.service, account: Keys.account)
+        resetPasscodeAttemptHistory()
+    }
+    func numberOfAttempts() -> Int {
+        return standardDefaults.integer(forKey: passcodeAttempts)
+    }
+    func resetPasscodeAttemptHistory() {
+        standardDefaults.removeObject(forKey: passcodeAttempts)
+    }
+    func recordIncorrectPasscodeAttempt() {
+        var numberOfAttemptsSoFar = standardDefaults.integer(forKey: passcodeAttempts)
+        numberOfAttemptsSoFar += 1
+        standardDefaults.set(passcodeAttempts, forKey: passcodeAttempts)
     }
 }
